Fixed about dependency non functional
Send method using export and require
dependency.server.test.js -> dependency.index.test.js

Send method using
dependency.route.test.js -> dependency.index.test.js

Send method using dependency
dependency.index.test.js -> dependency.server.test.js
